+++ index.js
@@ -56,6 +56,7 @@ function helmet (options) {
 helmet.contentSecurityPolicy = require('helmet-csp')
 helmet.dnsPrefetchControl = require('dns-prefetch-control')
 helmet.expectCt = require('expect-ct')
+helmet.featurePolicy = require('feature-policy')
 helmet.frameguard = require('frameguard')
 helmet.hidePoweredBy = require('hide-powered-by')
 helmet.hpkp = require('hpkp')