+++ HISTORY.md
@@ -1,3 +1,8 @@
+Unreleased
+==========
+
+- new: `featurePolicy` middleware
+
 3.13.0 / 2018-07-22
 ===================